Split out PackingService into a partial class with separate cs files for each service
