+++ Miscellany/ContainerPacking/PackingService/PackContainers.cs
@@ -1,0 +1,112 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using CromulentBisgetti.ContainerPacking.Entities; //added to interact with ContainerPacking
+using Autodesk.DesignScript.Runtime;
+
+namespace Miscellany.ContainerPacking
+{
+    /// <summary>
+    /// PackingService
+    /// </summary>
+    public static partial class PackingService
+    {
+        #region PackContainers
+
+        /// <summary>
+        /// Runs the chosen packing algorithm as a greedy strategy on a list of containers and a list of items to pack. It aims to solve optimally at each container in turn and is not globally optimal. The default algorithm (1) is EB-AFIT from David Chapman's 3DContainerPacking library.
+        /// </summary>
+        /// <param name="containers">Containers in order</param>
+        /// <param name="itemsToPack">Items to pack</param>
+        /// <param name="algorithm">Algorithm ID</param>
+        /// <returns name="packedItems">Items that were successfully packed</returns>
+        /// <returns name="unpackedItems">Items that were not packed</returns>
+        /// <returns name="isCompletePack">Are all items packed?</returns>
+        /// <returns name="packTimeInMilliseconds">Total pack time</returns>
+        /// <returns name="totalPackTimeInMilliseconds">Pack time per container</returns>
+        /// <returns name="percentContainerVolumePacked">Percentage of the container that is packed</returns>
+        /// <returns name="percentItemVolumePacked">Percentage of items packed</returns>
+        /// <search>
+        /// containerpacking
+        /// </search>
+        [MultiReturn(new[] { "packedItems", "unpackedItems", "isCompletePack", "packTimeInMilliseconds", "totalPackTimeInMilliseconds", "percentContainerVolumePacked", "percentItemVolumePacked" })]
+        public static Dictionary<string, object> PackContainers(List<Miscellany.ContainerPacking.Entities.Container> containers, List<Miscellany.ContainerPacking.Entities.Item> itemsToPack, int algorithm = 1)
+        {
+            //Select algorithm using integer
+            List<int> algorithms = new List<int> { algorithm };
+
+            //Create CromulentBisgetti Items
+            List<Item> items = new List<Item>();
+            foreach (Miscellany.ContainerPacking.Entities.Item i in itemsToPack)
+            {
+                decimal ddim1 = Miscellany.Math.Functions.ToDecimal(i.Dim1);
+                decimal ddim2 = Miscellany.Math.Functions.ToDecimal(i.Dim2);
+                decimal ddim3 = Miscellany.Math.Functions.ToDecimal(i.Dim3);
+                Item cbItem = new Item(i.ID, ddim1, ddim2, ddim3, i.Quantity);
+                items.Add(cbItem);
+            }
+
+            //Output lists
+            List<List<Miscellany.ContainerPacking.Entities.Item>> itemsPacked = new List<List<Miscellany.ContainerPacking.Entities.Item>>();
+            bool IsCompletePack = false;
+            List<int> PackTimeInMilliseconds = new List<int>();
+            int TotalPackTimeInMilliseconds = 0;
+            List<double> PercentContainerVolumePacked = new List<double>();
+            List<double> PercentItemVolumePacked = new List<double>();
+
+            //Loop through the containers
+            foreach (Miscellany.ContainerPacking.Entities.Container container in containers)
+            {
+                //Create CromulentBisgetti Container
+                decimal dLength = Miscellany.Math.Functions.ToDecimal(container.Length);
+                decimal dWidth = Miscellany.Math.Functions.ToDecimal(container.Width);
+                decimal dHeight = Miscellany.Math.Functions.ToDecimal(container.Height);
+                Container con = new Container(container.ID, dLength, dWidth, dHeight);
+                List<Container> cons = new List<Container> { con };
+                //Get container packing result
+                ContainerPackingResult containerPackingResult = CromulentBisgetti.ContainerPacking.PackingService.Pack(cons, items, algorithms).FirstOrDefault();
+                //Get the single algorthim packing result from the container packing result
+                AlgorithmPackingResult algorithmPackingResult = containerPackingResult.AlgorithmPackingResults.FirstOrDefault();
+                //Packed Items
+                List<Miscellany.ContainerPacking.Entities.Item> itemsPackedPass = new List<Miscellany.ContainerPacking.Entities.Item>();
+                foreach (Item i in algorithmPackingResult.PackedItems)
+                {
+                    Miscellany.ContainerPacking.Entities.Item mItem = ItemToMiscellany(i);
+                    itemsPackedPass.Add(mItem);
+                }
+                itemsPacked.Add(itemsPackedPass);
+                items = algorithmPackingResult.UnpackedItems; //items is set to unpacked items for next loop
+                IsCompletePack = algorithmPackingResult.IsCompletePack;
+                PackTimeInMilliseconds.Add(Convert.ToInt32(algorithmPackingResult.PackTimeInMilliseconds));
+                TotalPackTimeInMilliseconds += Convert.ToInt32(algorithmPackingResult.PackTimeInMilliseconds);
+                PercentContainerVolumePacked.Add(Miscellany.Math.Functions.ToDouble(algorithmPackingResult.PercentContainerVolumePacked));
+                PercentItemVolumePacked.Add(Miscellany.Math.Functions.ToDouble(algorithmPackingResult.PercentItemVolumePacked));
+                if (algorithmPackingResult.IsCompletePack == true)
+                {
+                    break;
+                }
+            }
+
+            //Convert CromulentBisgetti items to Miscellany Items for Unpacked Items
+            List<Miscellany.ContainerPacking.Entities.Item> itemsUnpacked = new List<Miscellany.ContainerPacking.Entities.Item>();
+            foreach (Item i in items)
+            {
+                Miscellany.ContainerPacking.Entities.Item mItem = ItemToMiscellany(i);
+                itemsUnpacked.Add(mItem);
+            }
+
+            //Return values
+            var d = new Dictionary<string, object>();
+            d.Add("packedItems", itemsPacked);
+            d.Add("unpackedItems", itemsUnpacked);
+            d.Add("isCompletePack", IsCompletePack);
+            d.Add("packTimeInMilliseconds", PackTimeInMilliseconds);
+            d.Add("totalPackTimeInMilliseconds", TotalPackTimeInMilliseconds);
+            d.Add("percentContainerVolumePacked", PercentContainerVolumePacked);
+            d.Add("percentItemVolumePacked", PercentItemVolumePacked);
+            return d;
+        }
+
+        #endregion
+    }
+}
