+++ Miscellany/Miscellany.csproj
@@ -72,7 +72,10 @@
     <Compile Include="About.cs" />
     <Compile Include="ContainerPacking\Entities\Container.cs" />
     <Compile Include="ContainerPacking\Entities\Item.cs" />
-    <Compile Include="ContainerPacking\PackingService.cs" />
+    <Compile Include="ContainerPacking\PackingService\PackContainer.cs" />
+    <Compile Include="ContainerPacking\PackingService\PackContainers.cs" />
+    <Compile Include="ContainerPacking\PackingService\PackContainersWithGroups.cs" />
+    <Compile Include="ContainerPacking\PackingService\PackContainersWithGroupsContinuously.cs" />
     <Compile Include="List\List.cs" />
     <Compile Include="Math\Functions.cs" />
     <Compile Include="Geometry\Abstract\Abstract.cs" />
